+++ ui/dashboard/src/components/stage/Chat/Messages.vue
@@ -7,13 +7,17 @@
   </div>
   <div v-else>
     <p v-for="item in messages" :key="item">
-      <small :style="'font-size: 1em'">
+      <small style="font-size: 1em">
         <b>{{ item.user }}: </b>
       </small>
       <span
         class="tag message"
         :style="
-          'background-color: ' + item.backgroundColor + '; color: ' + item.color + '; font-size: .85714em'
+          'background-color: ' +
+          item.backgroundColor +
+          '; color: ' +
+          item.color +
+          '; font-size: .85714em'
         "
         :class="messageClass[item.behavior]"
         :title="time(item.at)"
