+++ ui/dashboard/src/components/stage/Chat/Reaction.vue
@@ -7,20 +7,6 @@
   >
     {{ react }}
   </button>
-  <button
-    class="button is-small is-rounded reaction mx-1"
-    :key="encrease - fontSize"
-    @click="increateFontSize()"
-  >
-    {{ "➕" }}
-  </button>
-  <button
-    class="button is-small is-rounded reaction mx-1"
-    :key="decrease - fontSize"
-    @click="decreaseFontSize()"
-  >
-    {{ "➖" }}
-  </button>
   <span v-if="customEmoji">
     <ChatInput
       :picker-only="true"
@@ -56,7 +42,7 @@
 </template>
 
 <script>
-import { computed, reactive } from "vue";
+import { computed } from "vue";
 import { useStore } from "vuex";
 import anime from "animejs";
 import ChatInput from "@/components/form/ChatInput";
@@ -71,27 +57,6 @@ export default {
     const reactions = ["❤️", "🤣", "🙌", "👏"];
     const sendReaction = (react) => {
       store.dispatch("stage/sendReaction", react);
-    };
-
-    const fontSize = computed(() => store.state.stage.chat.fontSize);
-    const increateFontSize = () => {
-      let incValue = fontSize.value?.replace("px", "");
-      incValue++;
-      const parameters = reactive({
-        opacity: store.state.stage.chat.opacity,
-        fontSize: `${incValue}px`,
-      });
-      store.commit("stage/SET_CHAT_PARAMETERS", parameters);
-    };
-
-    const decreaseFontSize = () => {
-      let decValue = fontSize.value?.replace("px", "");
-      decValue > 1 && decValue--;
-      const parameters = reactive({
-        opacity: store.state.stage.chat.opacity,
-        fontSize: `${decValue}px`,
-      });
-      store.commit("stage/SET_CHAT_PARAMETERS", parameters);
     };
 
     const flyingReactions = computed(() => store.state.stage.reactions);
@@ -125,8 +90,6 @@ export default {
       flyin,
       flyout,
       sendCustomReaction,
-      increateFontSize,
-      decreaseFontSize,
     };
   },
 };
